MiniChamp Fix.
