+++ Scripts/Mobiles/Named/AcidElementalRenowned.cs
@@ -44,8 +44,6 @@ namespace Server.Mobiles
 
             this.VirtualArmor = 70;
 
-            this.PackItem(new EssenceSingularity());
-
             this.PackItem(new Nightshade(4));
             this.PackItem(new LesserPoisonPotion());
         }
