+++ Scripts/Mobiles/Named/DevourerRenowned.cs
@@ -45,8 +45,6 @@ namespace Server.Mobiles
 
             this.VirtualArmor = 44;
 
-            this.PackItem(new EssenceAchievement());
-
             this.PackNecroReg(24, 45);
         }
