+++ Scripts/Mobiles/Named/GreenGoblinAlchemistRenowned.cs
@@ -40,8 +40,6 @@ namespace Server.Mobiles
             this.Karma = -1500;
 
             this.VirtualArmor = 28;
-			
-            this.PackItem(new EssenceControl());
 
             switch ( Utility.Random(20) )
             {
