+++ Scripts/Mobiles/Named/PixieRenowned.cs
@@ -45,8 +45,6 @@ namespace Server.Mobiles
 
             this.VirtualArmor = 100;
 
-            this.PackItem(new EssenceFeeling());
-
             if (0.02 > Utility.RandomDouble())
                 this.PackStatue();				
         }
