+++ Scripts/Mobiles/Named/RakktaviRenowned.cs
@@ -44,8 +44,6 @@ namespace Server.Mobiles
             this.Karma = -6500;
 
             this.VirtualArmor = 56;
-
-            this.PackItem(new EssenceBalance());
 			
             this.AddItem(new Bow());
             this.PackItem(new Arrow(Utility.RandomMinMax(10, 30)));
