+++ Scripts/Mobiles/Named/WyvernRenowned.cs
@@ -46,8 +46,6 @@ namespace Server.Mobiles
             this.Karma = -24000;
 
             this.VirtualArmor = 70;
-
-            this.PackItem(new EssenceDiligence());
         }
 
         public WyvernRenowned(Serial serial)
