+++ Scripts/Services/MiniChampionSystem/OldSystem/MiniChampionAltar.cs
